Add Fake5Vault Task to prepare a environment variable. Revert changes in PaketRestore task
