+++ FAKE5.dev/package.json
@@ -3,17 +3,13 @@
     "tsc": "tsc"
   },
   "dependencies": {
-    "adm-zip": "^0.4.11",
     "semver": "^5.5.1",
-    "tmp": "0.0.31",
-    "typed-rest-client": "^1.0.9",
-    "vsts-fsharp-task-common": "file:../_build/vsts-fsharp-task-common-0.0.1.tgz",
-    "vsts-task-lib": "^2.0.2-preview",
-    "vsts-task-tool-lib": "^0.10.0"
+    "vsts-fsharp-task-common": "file:../_build/vsts-fsharp-task-common-0.1.0.tgz",
+    "vsts-task-lib": "^2.6.0"
   },
   "devDependencies": {
     "@types/node": "^7.0.12",
     "@types/semver": "^5.5.0",
-    "typescript": "^2.3.2"
+    "typescript": "^2.9.2"
   }
 }
