+++ FAKE5Vault.dev/package.json
@@ -3,13 +3,11 @@
     "tsc": "tsc"
   },
   "dependencies": {
-    "adm-zip": "^0.4.11",
-    "tmp": "0.0.31",
     "vsts-fsharp-task-common": "file:../_build/vsts-fsharp-task-common-0.1.0.tgz",
     "vsts-task-lib": "^2.6.0"
   },
   "devDependencies": {
-    "@types/node": "^7.0.69",
+    "@types/node": "^7.0.12",
     "typescript": "^2.9.2"
   }
 }
