+++ PaketRestore.dev/package.json
@@ -5,11 +5,11 @@
   "dependencies": {
     "adm-zip": "^0.4.11",
     "tmp": "0.0.31",
-    "vsts-fsharp-task-common": "file:../_build/vsts-fsharp-task-common-0.0.1.tgz",
+    "vsts-fsharp-task-common": "file:../_build/vsts-fsharp-task-common-0.1.0.tgz",
     "vsts-task-lib": "^2.0.2-preview"
   },
   "devDependencies": {
     "@types/node": "^7.0.12",
-    "typescript": "^2.3.2"
+    "typescript": "^2.9.2"
   }
 }
