+++ PaketRestore.dev/task.json
@@ -8,7 +8,7 @@
     "author": "Isaac Abraham",
     "version": {
         "Major": 0,
-        "Minor": 5,
+        "Minor": 4,
         "Patch": 0
     },
     "minimumAgentVersion": "1.83.0",    
@@ -32,8 +32,10 @@
         }
     ],
     "execution": {
-        "Node": {
-            "target": "runPaket.js"
+        "Process": {
+            "target": "$(currentDirectory)\\Paket.bat",
+            "argumentFormat": "$(PaketGroupName)",
+            "workingDirectory": "$(PaketPath)"
         }
     }
 }
