+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+# 1.3.0
+
+* Added a native FAKE 5 runner tasks
+* Added a task for using the FAKE 5 vault.
+
 # 1.2.0
 
 * Credential Provider only returns token for TFS/VSTS Links (not everyone)
