+++ build.fsx
@@ -206,7 +206,9 @@ module Node =
 
 open Fake.Core
 open Fake.Core.TargetOperators
-let dirs = ["PaketCredentialCleanup";"SetPaketCredentialProvider";"FSharpScript"; "FAKE4Runner"; "PaketRestore"; "FAKE5"]
+let dirs =
+    [ "PaketCredentialCleanup";"SetPaketCredentialProvider";
+      "FSharpScript"; "FAKE4Runner"; "PaketRestore"; "FAKE5"; "FAKE5Vault" ]
 let asDevel d = 
     let devDir = d + ".dev"
     if Directory.Exists devDir then devDir else d
