+++ vss-extension.json
@@ -1,7 +1,7 @@
 {
     "manifestVersion": 1,
     "id": "fsharp-helpers-extension",
-    "version": "1.2.0",
+    "version": "1.3.0",
     "name": "F# Helpers for VSTS",
     "publisher": "IsaacAbraham",
     "targets": [ { "id": "Microsoft.VisualStudio.Services" } ],
@@ -24,6 +24,7 @@
     },
     "files": [
         { "path": "FAKE5" },
+        { "path": "FAKE5Vault" },
         { "path": "FAKE4Runner" },
         { "path": "PaketRestore" },
         { "path": "FSharpScript" },
@@ -31,6 +32,11 @@
         { "path": "PaketCredentialCleanup" }
     ],
     "contributions": [
+        {   "id": "26d2a628-d5fe-4d5a-943d-33c78b2d76f3",
+            "type": "ms.vss-distributed-task.task",
+            "targets": [ "ms.vss-distributed-task.tasks" ],
+            "properties": { "name": "FAKE5Vault" }
+        },
         {   "id": "a2dadf20-1a83-4220-a4ee-b52f6c77f3cf",
             "type": "ms.vss-distributed-task.task",
             "targets": [ "ms.vss-distributed-task.tasks" ],
